backend/requirements.txt
django-cors-headers
django
psycopg2-binary
djoser
channels
channels-redis
django-rest-framework
celery==5.2.7
moviepy==1.0.3
yadisk==1.3.3